Basic tests for fileClient
