+++ pkg/logcli/client/client.go
@@ -1,11 +1,8 @@
 package client
 
 import (
-	"bufio"
-	"context"
 	"encoding/base64"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"net/http"
@@ -14,21 +11,14 @@ import (
 	"strings"
 	"time"
 
-	"github.com/cespare/xxhash"
 	"github.com/gorilla/websocket"
 	json "github.com/json-iterator/go"
 	"github.com/prometheus/common/config"
-	"github.com/prometheus/prometheus/pkg/labels"
-	"github.com/weaveworks/common/user"
 
-	"github.com/grafana/loki/pkg/iter"
 	"github.com/grafana/loki/pkg/loghttp"
 	"github.com/grafana/loki/pkg/logproto"
-	"github.com/grafana/loki/pkg/logql"
-	logqllog "github.com/grafana/loki/pkg/logql/log"
 	"github.com/grafana/loki/pkg/util"
 	"github.com/grafana/loki/pkg/util/build"
-	"github.com/grafana/loki/pkg/util/marshal"
 )
 
 const (
@@ -51,271 +41,6 @@ type Client interface {
 	Series(matchers []string, start, end time.Time, quiet bool) (*loghttp.SeriesResponse, error)
 	LiveTailQueryConn(queryStr string, delayFor time.Duration, limit int, start time.Time, quiet bool) (*websocket.Conn, error)
 	GetOrgID() string
-}
-
-type limiter struct {
-	n int
-}
-
-func (l *limiter) MaxQuerySeries(userID string) int {
-	return l.n
-}
-
-type querier struct {
-	r io.Reader
-}
-
-func (q *querier) SelectLogs(ctx context.Context, params logql.SelectLogParams) (iter.EntryIterator, error) {
-	expr, err := params.LogSelector()
-	if err != nil {
-		panic(err)
-	}
-	pipeline, err := expr.Pipeline()
-	if err != nil {
-		panic(err)
-	}
-	streampipe := pipeline.ForStream(labels.Labels{
-		labels.Label{Name: "foo", Value: "bar"},
-	})
-	it := NewFileIterator(q.r, "source:logcli", streampipe)
-
-	return it, nil
-}
-
-func (q *querier) SelectSamples(ctx context.Context, params logql.SelectSampleParams) (iter.SampleIterator, error) {
-	expr, err := params.Expr()
-	if err != nil {
-		panic(err)
-	}
-
-	sampleExtractor, err := expr.Extractor()
-	if err != nil {
-		panic(err)
-	}
-
-	streamSample := sampleExtractor.ForStream(labels.Labels{
-		labels.Label{Name: "foo", Value: "bar"},
-	})
-
-	it := NewFileSampleIterator(q.r, params.Start, params.End, "source:logcli", streamSample)
-
-	return it, nil
-}
-
-type FileSampleIterator struct {
-	s          *bufio.Scanner
-	labels     string
-	err        error
-	sp         logqllog.StreamSampleExtractor
-	curr       logproto.Sample
-	start, end time.Time
-}
-
-func NewFileSampleIterator(r io.Reader, start, end time.Time, labels string, sp logqllog.StreamSampleExtractor) *FileSampleIterator {
-	s := bufio.NewScanner(r)
-	s.Split(bufio.ScanLines)
-	return &FileSampleIterator{
-		s:      s,
-		labels: labels,
-		sp:     sp,
-		start:  start,
-		end:    end,
-	}
-}
-
-func (f *FileSampleIterator) Next() bool {
-	for f.s.Scan() {
-		value, _, ok := f.sp.Process([]byte(f.s.Text()))
-		ts := f.start.Add(2 * time.Minute)
-		if ok {
-			f.curr = logproto.Sample{
-				Timestamp: ts.UnixNano(),
-				Value:     value,
-				Hash:      xxhash.Sum64(f.s.Bytes()),
-			}
-			return true
-		}
-	}
-	return false
-}
-
-func (f *FileSampleIterator) Sample() logproto.Sample {
-	return f.curr
-}
-
-func (f *FileSampleIterator) Labels() string {
-	return f.labels
-}
-
-func (f *FileSampleIterator) Error() error {
-	return f.err
-}
-
-func (f *FileSampleIterator) Close() error {
-	// TODO: accept io.ReadCloser()?
-	return nil
-}
-
-type FileIterator struct {
-	s      *bufio.Scanner
-	labels string
-	err    error
-	sp     logqllog.StreamPipeline
-	curr   logproto.Entry
-}
-
-func NewFileIterator(r io.Reader, labels string, sp logqllog.StreamPipeline) *FileIterator {
-	s := bufio.NewScanner(r)
-	s.Split(bufio.ScanLines)
-	return &FileIterator{
-		s:      s,
-		labels: labels,
-		sp:     sp,
-	}
-}
-
-func (f *FileIterator) Next() bool {
-	for f.s.Scan() {
-		line, _, ok := f.sp.Process([]byte(f.s.Text()))
-		if ok {
-			f.curr = logproto.Entry{
-				Timestamp: time.Now(),
-				Line:      string(line),
-			}
-			return true
-		}
-	}
-	return false
-}
-
-func (f *FileIterator) Entry() logproto.Entry {
-	return f.curr
-}
-
-func (f *FileIterator) Labels() string {
-	return f.labels
-}
-
-func (f *FileIterator) Error() error {
-	return f.err
-}
-
-func (f *FileIterator) Close() error {
-	return nil
-}
-
-type FileClient struct {
-	r           io.Reader
-	labels      []string
-	labelValues []string
-	orgID       string
-	engine      *logql.Engine
-}
-
-func NewFileClient(r io.Reader) *FileClient {
-	eng := logql.NewEngine(logql.EngineOpts{}, &querier{r: r}, &limiter{n: 1024})
-	return &FileClient{
-		r:           r,
-		orgID:       "fake",
-		engine:      eng,
-		labels:      []string{"foo"},
-		labelValues: []string{"bar"},
-	}
-
-}
-
-func (f *FileClient) Query(q string, limit int, t time.Time, direction logproto.Direction, quiet bool) (*loghttp.QueryResponse, error) {
-	ctx := context.Background()
-
-	ctx = user.InjectOrgID(ctx, "fake")
-
-	params := logql.NewLiteralParams(
-		q,
-		t, t,
-		0,
-		0,
-		direction,
-		uint32(limit),
-		nil,
-	)
-
-	query := f.engine.Query(params)
-
-	result, err := query.Exec(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to exec query: %w", err)
-	}
-
-	value, err := marshal.NewResultValue(result.Data)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal result data: %w", err)
-	}
-
-	return &loghttp.QueryResponse{
-		Status: "success",
-		Data: loghttp.QueryResponseData{
-			ResultType: value.Type(),
-			Result:     value,
-			Statistics: result.Statistics,
-		},
-	}, nil
-}
-
-func (f *FileClient) QueryRange(queryStr string, limit int, start, end time.Time, direction logproto.Direction, step, interval time.Duration, quiet bool) (*loghttp.QueryResponse, error) {
-	ctx := context.Background()
-
-	ctx = user.InjectOrgID(ctx, "fake")
-
-	params := logql.NewLiteralParams(
-		queryStr,
-		start,
-		end,
-		step,
-		interval,
-		direction,
-		uint32(limit),
-		nil,
-	)
-
-	query := f.engine.Query(params)
-
-	result, err := query.Exec(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	value, err := marshal.NewResultValue(result.Data)
-	if err != nil {
-		return nil, err
-	}
-
-	return &loghttp.QueryResponse{
-		Status: "success",
-		Data: loghttp.QueryResponseData{
-			ResultType: value.Type(),
-			Result:     value,
-			Statistics: result.Statistics,
-		},
-	}, nil
-}
-
-func (f *FileClient) ListLabelNames(quiet bool, start, end time.Time) (*loghttp.LabelResponse, error) {
-	return &loghttp.LabelResponse{}, nil
-}
-
-func (f *FileClient) ListLabelValues(name string, quiet bool, start, end time.Time) (*loghttp.LabelResponse, error) {
-	return &loghttp.LabelResponse{}, nil
-}
-
-func (f *FileClient) Series(matchers []string, start, end time.Time, quiet bool) (*loghttp.SeriesResponse, error) {
-	return &loghttp.SeriesResponse{}, nil
-}
-func (f *FileClient) LiveTailQueryConn(queryStr string, delayFor time.Duration, limit int, start time.Time, quiet bool) (*websocket.Conn, error) {
-	panic("Not supported")
-}
-
-func (f *FileClient) GetOrgID() string {
-	return f.orgID
 }
 
 // Tripperware can wrap a roundtripper.
